# Api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Voting service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/learning/add_word": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new word to the vocabulary for the current language",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Add a new word",
                "parameters": [
                    {
                        "description": "New word details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/__.AddNewWordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.AddNewWordResp"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/all_languages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of all languages available for learning",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Get all available languages",
                "responses": {
                    "200": {
                        "description": "List of languages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/__.GetAllLanguagesResp"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/complete_lesson/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Complete the current lesson and update progress",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Complete a lesson",
                "parameters": [
                    {
                        "description": "Lesson completion details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/__.CompleteLessonReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.CompleteLessonResp"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/list_vocabulary/{language_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of all vocabulary for the current language",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Get all vocabulary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language ID",
                        "name": "language_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of vocabulary",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/__.Vocabulary"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/participate_lesson/{lesson_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Participate in an ongoing lesson and retrieve lesson details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Participate in a lesson",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "lesson_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.ParticipateLessonResp"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/start_exercise/{exercise_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Start a new exercise for the current lesson",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Start a new exercise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exercise ID",
                        "name": "exercise_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.StartExerciceResp"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/start_lesson": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Start a new lesson for the selected language",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Start a new lesson",
                "parameters": [
                    {
                        "description": "Lesson start details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/__.StartLearningReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.StartLearningResp"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/learning/submit_answer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit an answer for the current exercise",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learning"
                ],
                "summary": "Submit an answer for an exercise",
                "parameters": [
                    {
                        "description": "Answer submission details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/__.AnswerExerciseReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.AnswerExerciseResp"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "__.AddNewWordReq": {
            "type": "object",
            "properties": {
                "vocabulary": {
                    "$ref": "#/definitions/__.Vocabulary"
                }
            }
        },
        "__.AddNewWordResp": {
            "type": "object",
            "properties": {
                "language_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "word_id": {
                    "type": "string"
                }
            }
        },
        "__.AnswerExerciseReq": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "exercise_id": {
                    "type": "string"
                }
            }
        },
        "__.AnswerExerciseResp": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "earned_exp": {
                    "type": "integer"
                },
                "isCorrect": {
                    "type": "boolean"
                }
            }
        },
...
        "__.CompleteLessonReq": {
            "type": "object",
            "properties": {
                "lesson_id": {
                    "type": "string"
                }
            }
        },
        "__.CompleteLessonResp": {
            "type": "object",
            "properties": {
                "exp_earned": {
                    "type": "integer"
                },
                "next_lesson": {
                    "$ref": "#/definitions/__.Lessons"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "__.Exercise": {
            "type": "object",
            "properties": {
                "correct_answer": {
                    "type": "string"
                },
                "exercise_id": {
                    "type": "string"
                },
                "questions": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "__.GetAllLanguagesResp": {
            "type": "object",
            "properties": {
                "languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/__.Languages"
                    }
                }
            }
        },
...
        "__.Languages": {
            "type": "object",
            "properties": {
                "flag_emoji": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                },
                "language_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "__.Lessons": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "lesson_id": {
                    "type": "string"
                },
                "lesson_title": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                }
            }
        },
...
        "__.ParticipateLessonResp": {
            "type": "object",
            "properties": {
                "language": {
                    "$ref": "#/definitions/__.Languages"
                },
                "lesson": {
                    "$ref": "#/definitions/__.Lessons"
                },
                "theme": {
                    "type": "string"
                },
                "vocabulary": {
                    "$ref": "#/definitions/__.Vocabulary"
                }
            }
        },
...
        "__.StartExerciceResp": {
            "type": "object",
            "properties": {
                "exercise": {
                    "$ref": "#/definitions/__.Exercise"
                },
                "language": {
                    "$ref": "#/definitions/__.Languages"
                },
                "lesson": {
                    "$ref": "#/definitions/__.Lessons"
                }
            }
        },
        "__.StartLearningReq": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "__.StartLearningResp": {
            "type": "object",
            "properties": {
                "lesson_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "next_lesson": {
                    "$ref": "#/definitions/__.Lessons"
                }
            }
        },
...
        "__.Vocabulary": {
            "type": "object",
            "properties": {
                "example_lessons": {
                    "type": "string"
                },
                "translation": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authourization",
            "in": "header"
        }
    }
}